Preserve native absence storage and exact selected carriers

+++ test/cli-build/core/arrays/source-owned.test.mjs
@@ -43,11 +43,11 @@ test("CLI runs inferred source-owned array returns through finalized carrier fac
   assert.equal(build.status, 0, build.stdout + build.stderr);
 
   const generatedSource = await readFile(resolve(projectDirectory, "out/csharp/src/Index.cs"), "utf8");
-  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<double> make\(int value\)/);
-  assert.match(generatedSource, /return Tsonic\.CSharp\.Js\.JSArray<double>\.of\(\[value, value \+ 1\]\);/);
-  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<Tsonic\.CSharp\.Js\.JSArray<double>> nested\(int value\)/);
-  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<double> values/);
-  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<Tsonic\.CSharp\.Js\.JSArray<double>> nestedValues/);
+  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<int> make\(int value\)/);
+  assert.match(generatedSource, /return Tsonic\.CSharp\.Js\.JSArray<int>\.of\(\[value, value \+ 1\]\);/);
+  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<Tsonic\.CSharp\.Js\.JSArray<int>> nested\(int value\)/);
+  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<int> values/);
+  assert.match(generatedSource, /public static Tsonic\.CSharp\.Js\.JSArray<Tsonic\.CSharp\.Js\.JSArray<int>> nestedValues/);
   assert.match(generatedSource, /values\.length/);
   assert.doesNotMatch(generatedSource, /__unsupported|InvalidExpression|dynamic|System\.Reflection/);
