+++ test/cli-build/core/control-flow/statements.test.mjs
@@ -287,7 +287,7 @@ test("CLI routes top-level for-of statements through the C# module entrypoint", 
   assert.match(generatedSource, /total = 0;/);
   assert.match(generatedEntrypoint, /public static void Main\(\)/);
   assert.match(generatedEntrypoint, /Index\.__tsonic_module_init\(\);/);
-  assert.match(generatedSource, /foreach \(double value in new double\[\] \{ 1, 2, 3 \}\)/);
+  assert.match(generatedSource, /foreach \(int value in new int\[\] \{ 1, 2, 3 \}\)/);
   assert.match(generatedSource, /total = total \+ value;/);
   assert.doesNotMatch(generatedSource, /Top-level statement is outside/);
   assert.doesNotMatch(generatedSource, /__unsupported/);
